Make fancy things with FontAwesome

+++ assets/src/js/components/SortableImage.js
@@ -1,6 +1,8 @@
 'use strict'
 
 import React, { Component, PropTypes } from 'react'
+import FontAwesome from 'react-fontawesome'
+import { Button, Tooltip, OverlayTrigger } from 'react-bootstrap'
 
 export default class SortableImageList extends Component {
   constructor (props) {
@@ -18,11 +20,15 @@ export default class SortableImageList extends Component {
   delete (e) {
     e.preventDefault()
     const { id, onDelete } = this.props
-    onDelete(id)
+    if (confirm('Are you sure?')) onDelete(id)
   }
 
   render () {
-    const { filename, caption } = this.props
+    const { id, filename, caption } = this.props
+
+    const tooltip = (
+      <Tooltip id={`tooltip-delete-gallery-image-${id}`}>Delete</Tooltip>
+    )
 
     return (
       <li id={this.props.id}
@@ -35,11 +41,17 @@ export default class SortableImageList extends Component {
           placeholder='Caption'
           className='form-control gallery-image-input-caption'
           onBlur={this.update.bind(this)} />
-        <button
-          className='btn btn-default'
-          onClick={this.delete.bind(this)}
-          style={{marginTop: '-3.5px'}}
-          type='button'>Delete</button>
+        <OverlayTrigger placement='top' overlay={tooltip}>
+          <button
+            className='btn btn-default'
+            onClick={this.delete.bind(this)}
+            style={{marginTop: '-3.5px'}}
+            type='button'>
+            <FontAwesome
+              style={{color: 'red'}}
+              name='times' />
+          </button>
+        </OverlayTrigger>
       </li>
     )
   }
